refactored lvl_script
split off commands from the rest of lvl_script + a lib for all functions that are used by both the commands and rest of lvl_script + a file to store the volatile part that was split of and therefore doesn't belong in the lib
